+++ src/locales/ko/messages.po
@@ -2195,4 +2195,3 @@ msgstr ""
 
 msgid "Blink Validator"
 msgstr ""
-
