+++ src/locales/zh-Hans/messages.po
@@ -2195,4 +2195,3 @@ msgstr "× {0}"
 
 msgid "Blink Validator"
 msgstr "Blink 验证器"
-
